MEDICAL MATRIMONY
Medical Profession is one of the most arduous profession and in this journey of committing to the profession, doctors in our country tend to enter the marital bliss late in their lives as compared to their peers. Further, owing to their laborious schedule and long working hours they tend to give their personal lives a back seat and hence for a successful marriage it is very important to be with someone who understands the demands of the profession. This is the reason most doctors prefer to marry within their profession.

Looking at the above-mentioned challenges, Medical Dialogues bring to you Medical Matrimony (A unit of Everything Medical Treatment Pvt. Ltd.) (www.medicalmatrimony.in). Backed up by the huge elite readership base of Medical Dialogues ( over 8 lakhs medical practitioners across the country) Medical Matrimony gives medical professionals a platform to help them find like-minded people.

WEBSITE: www.medicalmatrimony.in