MEDICAL JOBS
Medicaljob.in is a one stop platform for all hiring and job related information in the medical and health sector of the country. It is one of its kind dedicated platform for healthcare jobseekers and recruiters to meet.

Medical Sector in the country, today suffers from strange problem of mismatch. While there is an extreme reported shortage of qualified healthcare professionals including doctors, nurses and other paramedical staff in the country, fresh passouts on the other hand are unable to find jobs due to lack of placement mechanisms in the healthcare sector.

To tackle this drawback, Medicaljob.in provides one stop junction to facilitate interaction between job seekers and recruiters.